bug fixing

+++ CMNetTrainer.py
@@ -239,8 +239,8 @@ class CMNetTrainer:
                                                 summary_op=tf.merge_summary(summary_ops))
 
     def roic2bbox(self, coord):
-        scale = tf.to_float(self.pre_processor.target_shape[:2] / self.model.sc_factor)
-        coord = tf.to_float(coord/scale)
+        scale = self.pre_processor.target_shape[:2] / self.model.sc_factor
+        coord = tf.to_float(coord)/scale
         bbox = tf.tile(coord, [1, 1, 2])
         bbox += [-0.02, -0.02, 0.02, 0.02]
         return bbox
